Refactoring and test fixing, disabling further coverage lines. To be improved.
