+++ webcomponents/search-webcomponent/src/app/data/store/marketplace.actions.ts
@@ -1,4 +1,4 @@
-import {FormFieldBase} from './../../shared/form/form-field-base';
+import {FormFieldBase} from '../../search-form/form/form-field-base';
 import {Category} from '../models/Category';
 import {createAction, props} from '@ngrx/store';
