+++ webcomponents/search-webcomponent/src/app/search-form/search-form.component.ts
@@ -1,8 +1,8 @@
-import {FormFieldControlService} from './../shared/form/form-field-control.service';
-import {FormFieldBase} from './../shared/form/form-field-base';
+import {FormFieldControlService} from './form/form-field-control.service';
+import {FormFieldBase} from './form/form-field-base';
 import {Component, Input, OnInit} from '@angular/core';
 import {FormGroup} from '@angular/forms';
-import {ValidationMessages} from '../utils/form/ValidationMessages';
+import {ValidationMessages} from '../utils/validators/ValidationMessages';
 
 @Component({
   selector: 'mp-search-form',
@@ -18,9 +18,7 @@ export class SearchFormComponent implements OnInit {
   @Input() appLanguage: string;
   @Input() currentCategoryKey: string;
 
-  constructor(
-    private formFieldControlService: FormFieldControlService
-  ) {}
+  constructor(private formFieldControlService: FormFieldControlService) {}
 
   ngOnInit() {
     this.form = this.formFieldControlService.toFormGroup(
