--- webcomponents/search-webcomponent/src/app/utils/form/ValidationMessages.ts
@@ -1,5 +1,0 @@
-export type ValidationMessages<T> = {
-  [P in keyof T]?: {
-    [key: string]: string
-  }
-}
